feat: Update layout files for book list and store screens
This commit updates the layout files for the book list and store screens. In the `fragment_book_list.xml` file, the `RecyclerView` now has a vertical orientation and is constrained to the parent layout. In the `fragment_store.xml` file, a `TabLayout` and `ViewPager2` are added for better navigation, and the `RecyclerView` for the bag is removed.