Add join button functionality to MessagesFragment
This commit adds the functionality to join a group in the MessagesFragment. It includes the implementation of the joinGroup() method, which creates a new document in the "BookClubMember" collection with the necessary data. The join button is also wired up to call the joinGroup() method when clicked. This allows users to join a book club group and become a member.