Refactor Gradle build files and update layout files and Implement pdfviewer
This commit refactors the Gradle build files by adding the jcenter repository and updating the Android Gradle Plugin to version 7.0.4. It also updates the layout files for the library group item and the PDF viewer fragment, adjusting the dimensions of the book cover image view. These changes improve the build configuration and enhance the user interface of the app.