Refactor fragment_messages.xml to add members button
This commit adds a new ImageButton to the fragment_messages.xml layout file. The button is used to display the group members and has a size of 30dp. It is positioned at the top right corner of the screen with a margin of 50dp from the end. This change enhances the user experience by providing easy access to the group members.

Co-authored-by: [Author Name]