Refactor Purchase class to add transactionId and approvalStatus properties
This commit adds the transactionId and approvalStatus properties to the Purchase class in Purchase.java. The transactionId property is of type String and is used to store the transaction ID for the purchase. The approvalStatus property is of type boolean and is used to indicate whether the purchase has been approved. These changes enhance the functionality of the Purchase class.

Co-authored-by: [Author Name]